Ideas for next step for dotenv migration
